+++ Puzzel_src/components/ColoredBetweenAsterisk.tsx
