+++ Puzzel_src/components/ShuffledArrayComponent.tsx
