+++ Puzzel_src/screens/resultScreen/ResultScreen.tsx
